pyproject.toml
[tool.poetry]
name = "donation-project"
version = "0.1.0"
description = ""
authors = ["Dauletnazarr <daulet.nazar@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
pillow = "^11.2.1"
drf-yasg = "^1.21.10"
python-dotenv = "^1.1.0"
django-filter = "^25.1"
redis = "^5.2.1"
django-redis = "^5.4.0"
celery = "^5.5.1"
faker = "^37.1.0"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
